+++ api/cpp/cbindgen.rs
@@ -137,6 +137,9 @@ fn builtin_structs(path: &Path) -> anyhow::Result<()> {
                     let pri_type = match stringify!($pri_type) {
                         "usize" => "uintptr_t",
                         "crate::animations::Instant" => "uint64_t",
+                        // This shouldn't be accessed by the C++ anyway, just need to have the same ABI in a struct
+                        "Option<i32>" => "std::pair<int32_t, int32_t>",
+                        "Option<core::ops::Range<i32>>" => "std::tuple<int32_t, int32_t, int32_t>",
                         other => other,
                     };
                     writeln!(file, "    {} {};", pri_type, stringify!($pri_field))?;
@@ -767,13 +770,13 @@ fn gen_interpreter(
     dependencies: &mut Vec<PathBuf>,
 ) -> anyhow::Result<()> {
     let mut config = default_config();
-    // Avoid Value, just export ValueOpaque.
     config.export.exclude = IntoIterator::into_iter([
         "Value",
         "ValueType",
         "PropertyDescriptor",
         "Diagnostic",
         "PropertyDescriptor",
+        "Box",
     ])
     .map(String::from)
     .collect();
@@ -794,7 +797,7 @@ fn gen_interpreter(
         "StructIteratorOpaque",
         "ComponentInstance",
         "StructIteratorResult",
-        "ValueOpaque",
+        "Value",
         "StructOpaque",
         "ModelNotifyOpaque",
     ])
@@ -820,6 +823,7 @@ fn gen_interpreter(
                 using slint::interpreter::ValueType;
                 using slint::interpreter::PropertyDescriptor;
                 using slint::interpreter::Diagnostic;
+                template <typename T> using Box = T*;
             }",
         )
         .generate()