+++ api/rs/slint/lib.rs
@@ -289,7 +289,7 @@ pub mod platform {
     /// This module contains the [`femtovg_renderer::FemtoVGRenderer`] and related types.
     ///
     /// It is only enabled when the `renderer-femtovg` Slint feature is enabled.
-    #[cfg(feature = "renderer-femtovg")]
+    #[cfg(all(feature = "renderer-femtovg", not(target_os = "android")))]
     pub mod femtovg_renderer {
         pub use i_slint_renderer_femtovg::FemtoVGRenderer;
         pub use i_slint_renderer_femtovg::OpenGLInterface;