+++ internal/renderers/skia/opengl_surface.rs
@@ -231,8 +231,8 @@ impl OpenGLSurface {
         // the system doesn't support transparency, we want to fall back to a config that doesn't - better than not
         // rendering anything at all. So we don't want to limit the configurations we get to see early on.
         // Commented out due to https://github.com/rust-windowing/glutin/issues/1640
-        // #[cfg(target_os = "macos")]
-        // let config_template_builder = config_template_builder.with_transparency(true);
+        #[cfg(target_os = "macos")]
+        let config_template_builder = config_template_builder.with_transparency(true);
 
         // Upstream advises to use this only on Windows.
         #[cfg(target_family = "windows")]